+++ tapscript/tx.go
@@ -1,7 +1,6 @@
 package tapscript
 
 import (
-	"bytes"
 	"context"
 	"crypto/sha256"
 	"errors"
@@ -11,12 +10,9 @@ import (
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/davecgh/go-spew/spew"
-	"github.com/lightninglabs/lndclient"
 	"github.com/lightninglabs/taproot-assets/asset"
 	"github.com/lightninglabs/taproot-assets/commitment"
 	"github.com/lightninglabs/taproot-assets/mssmt"
-	"github.com/lightninglabs/taproot-assets/tappsbt"
-	"github.com/lightningnetwork/lnd/input"
 )
 
 var (
@@ -289,132 +285,4 @@ func InputScriptSpendSigHash(virtualTx *wire.MsgTx, input *asset.Asset,
 		sigHashes, sigHashType, virtualTxCopy, zeroIndex,
 		prevOutFetcher, *tapLeaf,
 	)
-}
-
-// CreateTaprootSignature creates a Taproot signature for the given asset input.
-// Depending on the fields set in the input, this will either create a key path
-// spend or a script path spend.
-func CreateTaprootSignature(vIn *tappsbt.VInput, virtualTx *wire.MsgTx,
-	idx int, txSigner Signer) (wire.TxWitness, error) {
-
-	// Before we even attempt to sign anything, we need to make sure all the
-	// input information we require is present.
-	if len(vIn.TaprootBip32Derivation) == 0 {
-		return nil, fmt.Errorf("missing input Taproot BIP-0032 " +
-			"derivation")
-	}
-
-	// Currently, we only support creating one signature per input.
-	//
-	// TODO(guggero): Should we support signing multiple paths at the same
-	// time? What are the performance and security implications?
-	if len(vIn.TaprootBip32Derivation) > 1 {
-		return nil, fmt.Errorf("unsupported multiple taproot " +
-			"BIP-0032 derivation info found, can only sign for " +
-			"one at a time")
-	}
-	if len(vIn.TaprootBip32Derivation[0].LeafHashes) > 1 {
-		return nil, fmt.Errorf("unsupported number of leaf hashes in " +
-			"taproot BIP-0032 derivation info, can only sign for " +
-			"one at a time")
-	}
-
-	derivation := vIn.Bip32Derivation[0]
-	trDerivation := vIn.TaprootBip32Derivation[0]
-
-	keyDesc, err := tappsbt.KeyDescFromBip32Derivation(derivation)
-	if err != nil {
-		return nil, fmt.Errorf("error identifying input asset key "+
-			"descriptor from BIP-0032 derivation: %w", err)
-	}
-
-	// Compute a virtual prevOut from the input asset for the signer.
-	prevOut, err := InputAssetPrevOut(*vIn.Asset())
-	if err != nil {
-		return nil, err
-	}
-
-	// Start with a default sign descriptor and the BIP-0086 sign method
-	// then adjust depending on the input parameters.
-	spendDesc := lndclient.SignDescriptor{
-		KeyDesc:    keyDesc,
-		SignMethod: input.TaprootKeySpendBIP0086SignMethod,
-		Output:     prevOut,
-		HashType:   vIn.SighashType,
-		InputIndex: idx,
-	}
-
-	// There are three possible signing cases: BIP-0086 key spend path, key
-	// spend path with a script root, and script spend path.
-	switch {
-	// If there is no merkle root, we're doing a BIP-0086 key spend.
-	case len(vIn.TaprootMerkleRoot) == 0:
-		// This is the default case, so we don't need to do anything.
-
-	// No leaf hash means we're not signing a specific script, so this is
-	// the key spend path with a script root.
-	case len(vIn.TaprootMerkleRoot) == sha256.Size &&
-		len(trDerivation.LeafHashes) == 0:
-
-		spendDesc.SignMethod = input.TaprootKeySpendSignMethod
-		spendDesc.TapTweak = vIn.TaprootMerkleRoot
-
-	// One leaf hash and a merkle root means we're signing a specific
-	// script. There can be other scripts in the tree, but we only support
-	// creating a signature for a single one at a time.
-	case len(vIn.TaprootMerkleRoot) == sha256.Size &&
-		len(trDerivation.LeafHashes) == 1:
-
-		// If we're supposed to be signing for a leaf hash, we also
-		// expect the leaf script that hashes to that hash in the
-		// appropriate field.
-		if len(vIn.TaprootLeafScript) != 1 {
-			return nil, fmt.Errorf("specified leaf hash in " +
-				"taproot BIP-0032 derivation but missing " +
-				"taproot leaf script")
-		}
-
-		leafScript := vIn.TaprootLeafScript[0]
-		leaf := txscript.TapLeaf{
-			LeafVersion: leafScript.LeafVersion,
-			Script:      leafScript.Script,
-		}
-		leafHash := leaf.TapHash()
-		if !bytes.Equal(leafHash[:], trDerivation.LeafHashes[0]) {
-			return nil, fmt.Errorf("specified leaf hash in " +
-				"taproot BIP-0032 derivation but " +
-				"corresponding taproot leaf script was not " +
-				"found")
-		}
-
-		spendDesc.SignMethod = input.TaprootScriptSpendSignMethod
-		spendDesc.WitnessScript = leafScript.Script
-
-	// Some invalid combination of fields was specified, it's not clear what
-	// we should do. So rather than fail later, let's return an explicit
-	// error here.
-	default:
-		return nil, fmt.Errorf("unable to determine signing method " +
-			"from virtual transaction packet")
-	}
-
-	sig, err := txSigner.SignVirtualTx(&spendDesc, virtualTx, prevOut)
-	if err != nil {
-		return nil, err
-	}
-
-	witness := wire.TxWitness{sig.Serialize()}
-	if vIn.SighashType != txscript.SigHashDefault {
-		witness[0] = append(witness[0], byte(vIn.SighashType))
-	}
-
-	// If this was a script spend, we also have to add the script itself and
-	// the control block to the witness, otherwise the verifier will reject
-	// the generated witness.
-	if spendDesc.SignMethod == input.TaprootScriptSpendSignMethod {
-		witness = append(witness, spendDesc.WitnessScript)
-		witness = append(witness, vIn.TaprootLeafScript[0].ControlBlock)
-	}
-
-	return witness, nil
 }
