+++ tapsend/send_test.go
@@ -1,4 +1,4 @@
-package tapscript_test
+package tapsend_test
 
 import (
 	"bytes"
@@ -24,6 +24,7 @@ import (
 	"github.com/lightninglabs/taproot-assets/proof"
 	"github.com/lightninglabs/taproot-assets/tappsbt"
 	"github.com/lightninglabs/taproot-assets/tapscript"
+	"github.com/lightninglabs/taproot-assets/tapsend"
 	"github.com/lightninglabs/taproot-assets/vm"
 	"github.com/lightningnetwork/lnd/keychain"
 	"github.com/stretchr/testify/require"
@@ -736,7 +737,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 			state.address1, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		checkPreparedOutputsNonInteractive(
@@ -755,7 +756,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 			state.address2, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		checkPreparedOutputsNonInteractive(
@@ -774,7 +775,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 			state.address2, state.asset2PrevID,
 			state, state.asset2InputAssets, true,
 		)
-		return tapscript.PrepareOutputAssets(context.Background(), pkt)
+		return tapsend.PrepareOutputAssets(context.Background(), pkt)
 	},
 	err: commitment.ErrInvalidScriptKey,
 }, {
@@ -786,7 +787,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 			state.address2, state.asset2PrevID,
 			state, state.asset2InputAssets, true,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		checkPreparedOutputsInteractive(
@@ -806,7 +807,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 			state.asset1CollectGroupPrevID,
 			state, state.asset1CollectGroupInputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		checkPreparedOutputsNonInteractive(
@@ -825,7 +826,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 			state.address2, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		return tapscript.PrepareOutputAssets(context.Background(), pkt)
+		return tapsend.PrepareOutputAssets(context.Background(), pkt)
 	},
 	err: commitment.ErrInvalidScriptKey,
 }, {
@@ -847,7 +848,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 
 		// We expect an error because an interactive output cannot be
 		// un-spendable.
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.ErrorIs(t, err, commitment.ErrInvalidScriptKey)
 
 		// We also need to update the slit root's type to one that can
@@ -855,7 +856,7 @@ var prepareOutputAssetsTestCases = []testCase{{
 		pkt.Outputs[0].Type = tappsbt.TypePassiveSplitRoot
 
 		// Now we shouldn't get an error anymore.
-		return tapscript.PrepareOutputAssets(context.Background(), pkt)
+		return tapsend.PrepareOutputAssets(context.Background(), pkt)
 	},
 	err: nil,
 }}
@@ -886,11 +887,11 @@ var signVirtualTransactionTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		pkt.Inputs[0].Asset().Genesis = state.genesis1collect
-		return tapscript.SignVirtualTransaction(
+		return tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 	},
@@ -905,13 +906,13 @@ var signVirtualTransactionTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		firstPrevID := pkt.Outputs[0].Asset.PrevWitnesses[0].PrevID
 		firstPrevID.OutPoint.Index = 1337
 
-		return tapscript.SignVirtualTransaction(
+		return tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 	},
@@ -927,11 +928,11 @@ var signVirtualTransactionTestCases = []testCase{{
 			state.asset1CollectGroupPrevID, state,
 			state.asset1CollectGroupInputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		unvalidatedAsset := pkt.Outputs[0].Asset.Copy()
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
@@ -952,11 +953,11 @@ var signVirtualTransactionTestCases = []testCase{{
 			state.asset1CollectGroupPrevID, state,
 			state.asset1CollectGroupInputAssets, true,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		unvalidatedAsset := pkt.Outputs[0].Asset.Copy()
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
@@ -976,11 +977,11 @@ var signVirtualTransactionTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, true,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		unvalidatedAsset := pkt.Outputs[0].Asset.Copy()
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
@@ -999,11 +1000,11 @@ var signVirtualTransactionTestCases = []testCase{{
 			state.address1, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		unvalidatedAsset := pkt.Outputs[0].Asset.Copy()
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
@@ -1024,11 +1025,11 @@ var signVirtualTransactionTestCases = []testCase{{
 			state.asset1CollectGroupPrevID, state,
 			state.asset1CollectGroupInputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
 
 		unvalidatedAsset := pkt.Outputs[0].Asset.Copy()
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
@@ -1080,10 +1081,10 @@ var createOutputCommitmentsTestCases = []testCase{{
 			AnchorOutputTapscriptSibling: testPreimage,
 		})
 
-		_, err := tapscript.CreateOutputCommitments(nil, pkt, nil)
+		_, err := tapsend.CreateOutputCommitments(nil, pkt, nil)
 		return err
 	},
-	err: tapscript.ErrInvalidAnchorInfo,
+	err: tapsend.ErrInvalidAnchorInfo,
 }, {
 	name: "missing input asset commitment",
 	f: func(t *testing.T) error {
@@ -1094,9 +1095,9 @@ var createOutputCommitmentsTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
@@ -1110,14 +1111,14 @@ var createOutputCommitmentsTestCases = []testCase{{
 		err = inputCommitment.Delete(senderCommitment)
 		require.NoError(t, err)
 
-		_, err = tapscript.CreateOutputCommitments(
+		_, err = tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		return err
 	},
-	err: tapscript.ErrMissingAssetCommitment,
+	err: tapsend.ErrMissingAssetCommitment,
 }, {
 	name: "missing input asset",
 	f: func(t *testing.T) error {
@@ -1128,9 +1129,9 @@ var createOutputCommitmentsTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
@@ -1147,14 +1148,14 @@ var createOutputCommitmentsTestCases = []testCase{{
 		err = inputCommitment.Upsert(senderCommitment)
 		require.NoError(t, err)
 
-		_, err = tapscript.CreateOutputCommitments(
+		_, err = tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		return err
 	},
-	err: tapscript.ErrMissingAssetCommitment,
+	err: tapsend.ErrMissingAssetCommitment,
 }, {
 	name: "non-interactive collectible with group key",
 	f: func(t *testing.T) error {
@@ -1166,15 +1167,15 @@ var createOutputCommitmentsTestCases = []testCase{{
 			state.asset1CollectGroupPrevID, state,
 			state.asset1CollectGroupInputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1CollectGroupTapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
@@ -1194,15 +1195,15 @@ var createOutputCommitmentsTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, true,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
@@ -1222,15 +1223,15 @@ var createOutputCommitmentsTestCases = []testCase{{
 			state.address1, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset2TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
@@ -1251,15 +1252,15 @@ var createOutputCommitmentsTestCases = []testCase{{
 			state.address2, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset2TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
@@ -1281,15 +1282,15 @@ var createOutputCommitmentsTestCases = []testCase{{
 			state.asset1CollectGroupPrevID, state,
 			state.asset1CollectGroupInputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1CollectGroupTapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
@@ -1328,32 +1329,32 @@ var updateTaprootOutputKeysTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		require.NoError(t, err)
 
-		btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+		btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 		require.NoError(t, err)
 
 		outputCommitments[0] = nil
 
-		_, err = tapscript.UpdateTaprootOutputKeys(
+		_, err = tapsend.UpdateTaprootOutputKeys(
 			btcPkt, pkt, outputCommitments,
 		)
 		return err
 	},
-	err: tapscript.ErrMissingTapCommitment,
+	err: tapsend.ErrMissingTapCommitment,
 }, {
 	name: "missing receiver commitment",
 	f: func(t *testing.T) error {
@@ -1364,32 +1365,32 @@ var updateTaprootOutputKeysTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		require.NoError(t, err)
 
-		btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+		btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 		require.NoError(t, err)
 
 		outputCommitments[1] = nil
 
-		_, err = tapscript.UpdateTaprootOutputKeys(
+		_, err = tapsend.UpdateTaprootOutputKeys(
 			btcPkt, pkt, outputCommitments,
 		)
 		return err
 	},
-	err: tapscript.ErrMissingTapCommitment,
+	err: tapsend.ErrMissingTapCommitment,
 }, {
 	name: "interactive collectible with group key",
 	f: func(t *testing.T) error {
@@ -1400,25 +1401,25 @@ var updateTaprootOutputKeysTestCases = []testCase{{
 			state.asset1CollectGroupPrevID, state,
 			state.asset1CollectGroupInputAssets, true,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1CollectGroupTapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		require.NoError(t, err)
 
-		btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+		btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 		require.NoError(t, err)
 
-		_, err = tapscript.UpdateTaprootOutputKeys(
+		_, err = tapsend.UpdateTaprootOutputKeys(
 			btcPkt, pkt, outputCommitments,
 		)
 		require.NoError(t, err)
@@ -1439,25 +1440,25 @@ var updateTaprootOutputKeysTestCases = []testCase{{
 			state.address1, state.asset1PrevID,
 			state, state.asset1InputAssets, true,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		require.NoError(t, err)
 
-		btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+		btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 		require.NoError(t, err)
 
-		_, err = tapscript.UpdateTaprootOutputKeys(
+		_, err = tapsend.UpdateTaprootOutputKeys(
 			btcPkt, pkt, outputCommitments,
 		)
 		require.NoError(t, err)
@@ -1478,25 +1479,25 @@ var updateTaprootOutputKeysTestCases = []testCase{{
 			state.address1, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset2TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		require.NoError(t, err)
 
-		btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+		btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 		require.NoError(t, err)
 
-		_, err = tapscript.UpdateTaprootOutputKeys(
+		_, err = tapsend.UpdateTaprootOutputKeys(
 			btcPkt, pkt, outputCommitments,
 		)
 		require.NoError(t, err)
@@ -1518,25 +1519,25 @@ var updateTaprootOutputKeysTestCases = []testCase{{
 			state.address2, state.asset2PrevID,
 			state, state.asset2InputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset2TapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		require.NoError(t, err)
 
-		btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+		btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 		require.NoError(t, err)
 
-		_, err = tapscript.UpdateTaprootOutputKeys(
+		_, err = tapsend.UpdateTaprootOutputKeys(
 			btcPkt, pkt, outputCommitments,
 		)
 		require.NoError(t, err)
@@ -1559,25 +1560,25 @@ var updateTaprootOutputKeysTestCases = []testCase{{
 			state.asset1CollectGroupPrevID, state,
 			state.asset1CollectGroupInputAssets, false,
 		)
-		err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+		err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 		require.NoError(t, err)
-		err = tapscript.SignVirtualTransaction(
+		err = tapsend.SignVirtualTransaction(
 			pkt, state.signer, state.witnessValidator,
 		)
 		require.NoError(t, err)
 
 		inputCommitment := &state.asset1CollectGroupTapTree
-		outputCommitments, err := tapscript.CreateOutputCommitments(
+		outputCommitments, err := tapsend.CreateOutputCommitments(
 			tappsbt.InputCommitments{
 				0: inputCommitment,
 			}, pkt, nil,
 		)
 		require.NoError(t, err)
 
-		btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+		btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 		require.NoError(t, err)
 
-		_, err = tapscript.UpdateTaprootOutputKeys(
+		_, err = tapsend.UpdateTaprootOutputKeys(
 			btcPkt, pkt, outputCommitments,
 		)
 		require.NoError(t, err)
@@ -1612,25 +1613,25 @@ func createSpend(t *testing.T, state *spendData, inputSet commitment.InputSet,
 	// correct outputs.
 	pkt.Outputs[1].AnchorOutputIndex = 1
 
-	err := tapscript.PrepareOutputAssets(context.Background(), pkt)
+	err := tapsend.PrepareOutputAssets(context.Background(), pkt)
 	require.NoError(t, err)
-	err = tapscript.SignVirtualTransaction(
+	err = tapsend.SignVirtualTransaction(
 		pkt, state.signer, state.witnessValidator,
 	)
 	require.NoError(t, err)
 
 	inputCommitment := &state.asset2TapTree
-	outputCommitments, err := tapscript.CreateOutputCommitments(
+	outputCommitments, err := tapsend.CreateOutputCommitments(
 		tappsbt.InputCommitments{
 			0: inputCommitment,
 		}, pkt, nil,
 	)
 	require.NoError(t, err)
 
-	btcPkt, err := tapscript.CreateAnchorTx(pkt.Outputs)
+	btcPkt, err := tapsend.CreateAnchorTx(pkt.Outputs)
 	require.NoError(t, err)
 
-	_, err = tapscript.UpdateTaprootOutputKeys(
+	_, err = tapsend.UpdateTaprootOutputKeys(
 		btcPkt, pkt, outputCommitments,
 	)
 	require.NoError(t, err)
@@ -1873,9 +1874,9 @@ func TestAreValidAnchorOutputIndexes(t *testing.T) {
 	outputs := []*tappsbt.VOutput{}
 
 	// Reject groups of outputs smaller than 1.
-	assetOnlySpend, err := tapscript.AreValidAnchorOutputIndexes(outputs)
+	assetOnlySpend, err := tapsend.AreValidAnchorOutputIndexes(outputs)
 	require.False(t, assetOnlySpend)
-	require.ErrorIs(t, err, tapscript.ErrInvalidOutputIndexes)
+	require.ErrorIs(t, err, tapsend.ErrInvalidOutputIndexes)
 
 	// Insert a locator for the sender and for the receiver, that would form
 	// a Taproot Asset only spend.
@@ -1885,14 +1886,14 @@ func TestAreValidAnchorOutputIndexes(t *testing.T) {
 		AnchorOutputIndex: 1,
 	}}
 
-	assetOnlySpend, err = tapscript.AreValidAnchorOutputIndexes(outputs)
+	assetOnlySpend, err = tapsend.AreValidAnchorOutputIndexes(outputs)
 	require.True(t, assetOnlySpend)
 	require.NoError(t, err)
 
 	// Modify the receiver locator so the indexes are no longer continuous.
 	outputs[1].AnchorOutputIndex = 2
 
-	assetOnlySpend, err = tapscript.AreValidAnchorOutputIndexes(outputs)
+	assetOnlySpend, err = tapsend.AreValidAnchorOutputIndexes(outputs)
 	require.False(t, assetOnlySpend)
 	require.NoError(t, err)
 
@@ -1901,7 +1902,7 @@ func TestAreValidAnchorOutputIndexes(t *testing.T) {
 		AnchorOutputIndex: 1,
 	})
 
-	assetOnlySpend, err = tapscript.AreValidAnchorOutputIndexes(outputs)
+	assetOnlySpend, err = tapsend.AreValidAnchorOutputIndexes(outputs)
 	require.True(t, assetOnlySpend)
 	require.NoError(t, err)
 }
@@ -1927,8 +1928,8 @@ func TestAddressValidInput(t *testing.T) {
 	}
 }
 
-func addrToFundDesc(addr address.Tap) *tapscript.FundingDescriptor {
-	return &tapscript.FundingDescriptor{
+func addrToFundDesc(addr address.Tap) *tapsend.FundingDescriptor {
+	return &tapsend.FundingDescriptor{
 		ID:       addr.AssetID,
 		GroupKey: addr.GroupKey,
 		Amount:   addr.Amount,
@@ -1947,14 +1948,14 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 		state := initSpendScenario(t)
 		fundDesc := addrToFundDesc(state.address1)
 
-		inputAsset, err := tapscript.AssetFromTapCommitment(
+		inputAsset, err := tapsend.AssetFromTapCommitment(
 			&state.asset1TapTree, fundDesc, state.spenderScriptKey,
 		)
 		if err != nil {
 			return nil, nil, err
 		}
 
-		fullValue, err := tapscript.ValidateInputs(
+		fullValue, err := tapsend.ValidateInputs(
 			tappsbt.InputCommitments{
 				0: &state.asset1TapTree,
 			}, []*btcec.PublicKey{&state.spenderScriptKey},
@@ -1974,7 +1975,7 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 		state := initSpendScenario(t)
 		fundDesc := addrToFundDesc(state.address1CollectGroup)
 
-		inputAsset, err := tapscript.AssetFromTapCommitment(
+		inputAsset, err := tapsend.AssetFromTapCommitment(
 			&state.asset1CollectGroupTapTree, fundDesc,
 			state.spenderScriptKey,
 		)
@@ -1982,7 +1983,7 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 			return nil, nil, err
 		}
 
-		fullValue, err := tapscript.ValidateInputs(
+		fullValue, err := tapsend.ValidateInputs(
 			tappsbt.InputCommitments{
 				0: &state.asset1CollectGroupTapTree,
 			}, []*btcec.PublicKey{&state.spenderScriptKey},
@@ -2002,14 +2003,14 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 		state := initSpendScenario(t)
 		fundDesc := addrToFundDesc(state.address1)
 
-		inputAsset, err := tapscript.AssetFromTapCommitment(
+		inputAsset, err := tapsend.AssetFromTapCommitment(
 			&state.asset2TapTree, fundDesc, state.spenderScriptKey,
 		)
 		if err != nil {
 			return nil, nil, err
 		}
 
-		fullValue, err := tapscript.ValidateInputs(
+		fullValue, err := tapsend.ValidateInputs(
 			tappsbt.InputCommitments{
 				0: &state.asset2TapTree,
 			}, []*btcec.PublicKey{&state.spenderScriptKey},
@@ -2029,14 +2030,14 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 		state := initSpendScenario(t)
 		fundDesc := addrToFundDesc(state.address2)
 
-		inputAsset, err := tapscript.AssetFromTapCommitment(
+		inputAsset, err := tapsend.AssetFromTapCommitment(
 			&state.asset1TapTree, fundDesc, state.spenderScriptKey,
 		)
 		if err != nil {
 			return nil, nil, err
 		}
 
-		fullValue, err := tapscript.ValidateInputs(
+		fullValue, err := tapsend.ValidateInputs(
 			tappsbt.InputCommitments{
 				0: &state.asset1TapTree,
 			}, []*btcec.PublicKey{&state.spenderScriptKey},
@@ -2049,21 +2050,21 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 
 		return &state.asset1, inputAsset, nil
 	},
-	err: tapscript.ErrInsufficientInputAssets,
+	err: tapsend.ErrInsufficientInputAssets,
 }, {
 	name: "collectible with missing input asset",
 	f: func(t *testing.T) (*asset.Asset, *asset.Asset, error) {
 		state := initSpendScenario(t)
 		fundDesc := addrToFundDesc(state.address1CollectGroup)
 
-		inputAsset, err := tapscript.AssetFromTapCommitment(
+		inputAsset, err := tapsend.AssetFromTapCommitment(
 			&state.asset1TapTree, fundDesc, state.spenderScriptKey,
 		)
 		if err != nil {
 			return nil, nil, err
 		}
 
-		fullValue, err := tapscript.ValidateInputs(
+		fullValue, err := tapsend.ValidateInputs(
 			tappsbt.InputCommitments{
 				0: &state.asset1TapTree,
 			}, []*btcec.PublicKey{&state.spenderScriptKey},
@@ -2076,7 +2077,7 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 
 		return &state.asset1, inputAsset, nil
 	},
-	err: tapscript.ErrMissingInputAsset,
+	err: tapsend.ErrMissingInputAsset,
 }, {
 	name: "normal with bad sender script key",
 	f: func(t *testing.T) (*asset.Asset, *asset.Asset, error) {
@@ -2092,14 +2093,14 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 
 		fundDesc := addrToFundDesc(*address1testnet)
 
-		inputAsset, err := tapscript.AssetFromTapCommitment(
+		inputAsset, err := tapsend.AssetFromTapCommitment(
 			&state.asset1TapTree, fundDesc, state.receiverPubKey,
 		)
 		if err != nil {
 			return nil, nil, err
 		}
 
-		fullValue, err := tapscript.ValidateInputs(
+		fullValue, err := tapsend.ValidateInputs(
 			tappsbt.InputCommitments{
 				0: &state.asset1TapTree,
 			}, []*btcec.PublicKey{&state.spenderScriptKey},
@@ -2112,7 +2113,7 @@ var addressValidInputTestCases = []addressValidInputTestCase{{
 
 		return &state.asset1, inputAsset, nil
 	},
-	err: tapscript.ErrMissingInputAsset,
+	err: tapsend.ErrMissingInputAsset,
 }}
 
 // TestPayToAddrScript tests edge cases around creating a P2TR script with
